feat: structured code and added saving data
- divided code on several files and folders
- added saving scraped each scraped lead in local storage
- added reading data from local storage
