+++ public/manifest.json
@@ -1,7 +1,7 @@
 {
   "manifest_version": 3,
   "name": "Sales Navigator Demon",
-  "version": "1.4.1",
+  "version": "1.5.0",
   "description": "Extract publicly available lead data from LinkedIn Sales Navigator efficiently and export it in CSV format.",
   "permissions": ["activeTab", "scripting", "storage", "tabs"],
   "host_permissions": ["https://www.linkedin.com/sales/lists/people",
