+++ src/utils/constants.js
@@ -1,0 +1,11 @@
+export const headerArray = [
+  "FullName",
+  "LeadLocation",
+  "JobTitle",
+  "CompanyName",
+  "CompanyLocation",
+  "CompanyIndustry",
+  "CompanyWebsite"
+];
+
+export const baseLinkedinUrl = "https://www.linkedin.com";
